Deprecate postbox factory global
